+++ app/lib/makeArr.ts
@@ -1,0 +1,9 @@
+export default function makeArr(length: number) {
+  const result = [];
+
+  for (let i = 0; i < length; i++) {
+    result.push(i);
+  }
+
+  return result;
+}
