+++ app/ui/invoices/pagination/number.tsx
@@ -1,0 +1,35 @@
+import cn from "@/app/lib/cn";
+import Link from "next/link";
+
+export default function PaginationNumber({
+  page,
+  href,
+  isActive,
+  position,
+}: {
+  page: number | string;
+  href: string;
+  position?: "first" | "last" | "middle" | "single";
+  isActive: boolean;
+}) {
+  const className = cn(
+    "flex h-10 w-10 items-center justify-center text-sm border",
+    {
+      "rounded-l-md": position === "first" || position === "single",
+      "rounded-r-md": position === "last" || position === "single",
+      "z-10 bg-blue-600 border-blue-600 text-white": isActive,
+      "hover:bg-gray-100": !isActive && position !== "middle",
+      "text-gray-300": position === "middle",
+    }
+  );
+
+  if (isActive || position === "middle") {
+    return <div className={className}>{page}</div>;
+  }
+
+  return (
+    <Link href={href} className={className}>
+      {page}
+    </Link>
+  );
+}
