--- app/ui/invoices/pagination/pagination-content.tsx
@@ -1,129 +1,0 @@
-"use client";
-
-import cn from "@/app/lib/cn";
-import { generatePagination } from "@/app/lib/utils";
-import { ArrowLeftIcon, ArrowRightIcon } from "@heroicons/react/24/outline";
-import Link from "next/link";
-import { usePathname, useSearchParams } from "next/navigation";
-
-type Props = {
-  totalPages: number;
-};
-
-export default function PaginationContent({ totalPages }: Props) {
-  const searchParams = useSearchParams();
-  const pathname = usePathname();
-  const currPage = Number(searchParams.get("page") ?? 1);
-  const allPages = generatePagination(currPage, totalPages);
-
-  function createPageUrl(pageNumber: number | string) {
-    const params = new URLSearchParams(searchParams);
-    params.set("page", pageNumber.toString());
-    return `${pathname}?${params.toString()}`;
-  }
-
-  return (
-    <div className="inline-flex">
-      <PaginationArrow
-        direction="left"
-        href={createPageUrl(currPage - 1)}
-        isDisabled={currPage <= 1}
-      />
-
-      <div className="flex -space-x-px">
-        {allPages.map((page, index) => {
-          let position: "first" | "last" | "single" | "middle" | undefined;
-
-          if (index === 0) position = "first";
-          if (index === allPages.length - 1) position = "last";
-          if (allPages.length === 1) position = "single";
-          if (page === "...") position = "middle";
-
-          return (
-            <PaginationNumber
-              key={page}
-              href={createPageUrl(page)}
-              page={page}
-              position={position}
-              isActive={currPage === page}
-            />
-          );
-        })}
-      </div>
-
-      <PaginationArrow
-        direction="right"
-        href={createPageUrl(currPage + 1)}
-        isDisabled={currPage >= totalPages}
-      />
-    </div>
-  );
-}
-
-function PaginationNumber({
-  page,
-  href,
-  isActive,
-  position,
-}: {
-  page: number | string;
-  href: string;
-  position?: "first" | "last" | "middle" | "single";
-  isActive: boolean;
-}) {
-  const className = cn(
-    "flex h-10 w-10 items-center justify-center text-sm border",
-    {
-      "rounded-l-md": position === "first" || position === "single",
-      "rounded-r-md": position === "last" || position === "single",
-      "z-10 bg-blue-600 border-blue-600 text-white": isActive,
-      "hover:bg-gray-100": !isActive && position !== "middle",
-      "text-gray-300": position === "middle",
-    }
-  );
-
-  if (isActive || position === "middle") {
-    return <div className={className}>{page}</div>;
-  }
-
-  return (
-    <Link href={href} className={className}>
-      {page}
-    </Link>
-  );
-}
-
-function PaginationArrow({
-  href,
-  direction,
-  isDisabled,
-}: {
-  href: string;
-  direction: "left" | "right";
-  isDisabled?: boolean;
-}) {
-  const className = cn(
-    "flex h-10 w-10 items-center justify-center rounded-md border",
-    {
-      "pointer-events-none text-gray-300": isDisabled,
-      "hover:bg-gray-100": !isDisabled,
-      "mr-2 md:mr-4": direction === "left",
-      "ml-2 md:ml-4": direction === "right",
-    }
-  );
-
-  const icon =
-    direction === "left" ? (
-      <ArrowLeftIcon className="w-4" />
-    ) : (
-      <ArrowRightIcon className="w-4" />
-    );
-
-  return isDisabled ? (
-    <div className={className}>{icon}</div>
-  ) : (
-    <Link className={className} href={href}>
-      {icon}
-    </Link>
-  );
-}
