+++ app/ui/skeletons.tsx
@@ -1,3 +1,5 @@
+import PaginationArrow from "./invoices/pagination/arrow";
+
 // Loading animation
 const shimmer =
   "before:absolute before:inset-0 before:-translate-x-full before:animate-[shimmer_2s_infinite] before:bg-gradient-to-r before:from-transparent before:via-white/60 before:to-transparent";
@@ -213,6 +215,16 @@ export function InvoicesTableSkeleton() {
           </table>
         </div>
       </div>
+    </div>
+  );
+}
+
+export function PaginationSkeleton() {
+  return (
+    <div className="flex">
+      <PaginationArrow direction="left" />
+      <div></div>
+      <PaginationArrow direction="right" />
     </div>
   );
 }
